+++ src/components/Layout/Footer.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+const Footer = () => {
+  return (
+    <footer>
+      © {new Date().getFullYear()}, Built with
+      {` `}
+      <a href="https://www.gatsbyjs.com">Gatsby</a>
+    </footer>
+  );
+};
+
+export default Footer;
